Refactor Office.js mock into modular files
