+++ tests/_support/office-mock.ts
@@ -1,7 +1,15 @@
 import type { Page } from "@playwright/test";
+import fs from "node:fs/promises";
 import path from "node:path";
 
-const officeMockPath = path.join(process.cwd(), "tests", "_support", "browser-mocks", "office.ts");
+const officeMockDir = path.join(process.cwd(), "tests", "_support", "office");
+const officeMockFiles = [
+  "types.ts",
+  "office-primitives.ts",
+  "document-model.ts",
+  "adapter.ts",
+  "install.ts",
+];
 
 /** Replaces the hosted Office script with the minimal APIs needed for task pane startup. */
 export async function installOfficeMock(page: Page) {
@@ -11,6 +19,14 @@ export async function installOfficeMock(page: Page) {
 }
 
 async function compileOfficeMock() {
-  const { compileBrowserMock } = await import("./transpile-browser-mock");
-  return compileBrowserMock(officeMockPath);
+  const { compileBrowserMockSource } = await import("./transpile-browser-mock");
+  const sourceParts = await Promise.all(
+    officeMockFiles.map(fileName => fs.readFile(path.join(officeMockDir, fileName), "utf8")),
+  );
+
+  return compileBrowserMockSource(
+    officeMockDir,
+    sourceParts.join("\n\n"),
+    path.join(officeMockDir, "index.ts"),
+  );
 }
